Revert "MaxSpeed & fixes"
This reverts commit 93d8f7506901b9988371583e056d14c1b6341ee9.

+++ brickBreak/bar.h
@@ -10,7 +10,7 @@ public:
 	int isInsideScreen(sf::Vector2f&);
 	sf::Vector2f getDim() const;
 	sf::Vector2f getPos() const;
-	float getTanBar() const;
+	double getTanBar() const;
 	int setPosx(float);
 	int setPosy(float);
 	void draw(sf::RenderWindow&);
